Moved MiddleTopColumn to middleTopCol folder
