+++ src/components/AicureToolFull.tsx
@@ -2,13 +2,11 @@
 import { useState } from "react";
 import LeftColumn from "./LeftColumn";
 import RightColumn from "./RightColumn";
-import MiddleTopColumn from "./MiddleTopColumn";
+import MiddleTopColumn from "./middleTopCol/MiddleTopColumn";
 import MiddleBottomColumn from "./MiddleBottomColumn";
 import { useIsRightVisible } from "@/store/useIsRightVisible";
 
 export default function AicureToolFull() {
-  // const [showRight, setShowRight] = useState(true);
-  // const toggleRightColumn = () => setShowRight((prev) => !prev);
   const isRightColumnVisible = useIsRightVisible(
     (state) => state.isRightColumnVisible
   );
